INTYGFV-14827: Moved handling of key for getCertificatesForPatient into CertificateForPatientService.
